+++ frontend/src/components/Button/enums/ButtonTypes.ts
@@ -1,0 +1,6 @@
+export enum ButtonTypes {
+  Action = "action",
+  DangerousAction = "dangerous-action",
+  Default = "default",
+  Submit = "submit",
+}
