--- frontend/src/components/enums/ButtonTypes.ts
@@ -1,6 +1,0 @@
-export enum ButtonTypes {
-    Action = 'action',
-    DangerousAction = 'dangerous-action',
-    Default = 'default',
-    Submit = 'submit',
-  }
